utils for testing in module NefCore
