+++ project/Tests/CoreTests/JekyllTests.swift
@@ -2,40 +2,6 @@
 
 import XCTest
 @testable import NefCore
-import AppKit
-import Bow
-import BowEffects
-import NefModels
-
-extension XCTestCase {
-    func markdown(content: String) -> IO<CoreRenderError, RendererOutput<String>> {
-        CoreRender.markdown.render(content: content)
-                           .provide(CoreMarkdownEnvironment())
-    }
-    
-    func jekyll(content: String, permalink: String) -> IO<CoreRenderError, RendererOutput<String>> {
-        CoreRender.jekyll.render(content: content)
-                         .provide(CoreJekyllEnvironment(permalink: permalink))
-    }
-    
-    func carbon(content: String, downloader: CarbonDownloader, style: CarbonStyle) -> IO<CoreRenderError, RendererOutput<NSImage>> {
-        CoreRender.carbon.render(content: content)
-                         .provide(CoreCarbonEnvironment.init(downloader: downloader, style: style))
-    }
-}
-
-extension XCTestCase {
-    func assert<A: Equatable>(_ io: IO<CoreRenderError, RendererOutput<A>>, succeeds: A) {
-        _ = io.unsafeRunSyncEither().bimap({ e in XCTFail(e.localizedDescription) },
-                                           { rendered in XCTAssertEqual(rendered.output.all().first!, succeeds) })
-    }
-    
-    func assert<A: Equatable>(_ io: IO<CoreRenderError, RendererOutput<A>>, fails: CoreRenderError) {
-        _ = io.unsafeRunSyncEither().bimap({ e in XCTAssertEqual(e, fails) },
-                                           { rendered in XCTFail("\(rendered.output)") })
-    }
-}
-
 
 class JekyllTests: XCTestCase {
